28-05-2025

+++ JavaCodes/src/Rectangle.java
@@ -1,20 +1,20 @@
 
 public class Rectangle {
-    private int width;
-    private int length;
-    public Rectangle(int width,int length) {
+    private double width;
+    private double length;
+    public Rectangle(double width,double length) {
     	if(width<0)this.width=0;
     	else this.width=width;
     	if(length<0)this.length=0;
     	else this.length=length;
     }
-    public int getWidth() {
+    public double getWidth() {
     	return this.width;
     }
-    public int getLenth() {
+    public double getLenth() {
     	return this.length;
     }
-    public int getArea() {
+    public double getArea() {
     	return this.width*this.length;
     }
 }
